Extract duplicated TimeUtils utility
Extracted duplicate `formatRelativeTime` and `buildMetadata` functions,
along with their associated constants, into a shared `TimeUtils` object.
Updated all call sites in DashboardVoicesFragment, DashboardPostDetailActivity,
and PostDetailAdapter.
